+++ apps/web/src/components/task-list.tsx
@@ -1,27 +1,10 @@
 "use client";
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
-import {
-  CheckCircle2,
-  XCircle,
-  Pause,
-  Loader2,
-  Archive,
-  Plus,
-  Minus,
-} from "lucide-react";
+import { Archive } from "lucide-react";
 import { Badge } from "@/components/ui/badge";
+import { Task, type Task as TaskType } from "./task";
 
-interface Task {
-  id: string;
-  title: string;
-  status: "running" | "interrupted" | "done" | "error";
-  date: string;
-  repository: string;
-  additions?: number;
-  deletions?: number;
-}
-
-const tasks: Task[] = [
+const tasks: TaskType[] = [
   {
     id: "1",
     title: "Identify and propose code improvements",
@@ -53,65 +36,6 @@ const tasks: Task[] = [
     repository: "langchain-ai/open_deep_research",
   },
 ];
-
-const StatusIndicator = ({ status }: { status: Task["status"] }) => {
-  switch (status) {
-    case "running":
-      return <Loader2 className="h-3 w-3 animate-spin text-blue-500" />;
-    case "interrupted":
-      return <Pause className="h-3 w-3 text-yellow-500" />;
-    case "done":
-      return <CheckCircle2 className="h-3 w-3 text-green-500" />;
-    case "error":
-      return <XCircle className="h-3 w-3 text-red-500" />;
-    default:
-      return null;
-  }
-};
-
-const TaskItem = ({ task }: { task: Task }) => {
-  return (
-    <div className="group flex items-start gap-3 border-b border-gray-100 py-3 last:border-b-0">
-      <div className="mt-1 flex-shrink-0">
-        <StatusIndicator status={task.status} />
-      </div>
-
-      <div className="min-w-0 flex-1">
-        <div className="flex items-start justify-between">
-          <div className="flex-1">
-            <h3 className="mb-1 text-sm font-medium text-gray-900">
-              {task.title}
-            </h3>
-            <div className="flex items-center gap-2 text-xs text-gray-500">
-              <span>{task.date}</span>
-              <span>·</span>
-              <span>{task.repository}</span>
-            </div>
-          </div>
-
-          <div className="ml-4 flex items-center gap-2">
-            {task.status === "interrupted" && (
-              <Archive className="h-4 w-4 text-gray-400" />
-            )}
-
-            {task.additions !== undefined && task.deletions !== undefined && (
-              <div className="flex items-center gap-1 text-xs">
-                <span className="flex items-center gap-0.5 text-green-600">
-                  <Plus className="h-3 w-3" />
-                  {task.additions}
-                </span>
-                <span className="flex items-center gap-0.5 text-red-600">
-                  <Minus className="h-3 w-3" />
-                  {task.deletions}
-                </span>
-              </div>
-            )}
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-};
 
 export default function TaskList() {
   return (
@@ -151,7 +75,7 @@ export default function TaskList() {
         >
           <div className="space-y-0">
             {tasks.map((task) => (
-              <TaskItem
+              <Task
                 key={task.id}
                 task={task}
               />
